+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # JSONSchema Changelog
+
+## 0.4.0
+
+- Converted project to Swift 3
+- Adds `addFormat` method to `Schema` which supports custom format definitions
+- Renames `ValidationResult.Valid` to `ValidationResult.valid` to match Swift 3 naming conventions
+
 ## 0.3.0
 ### Enhancements